no longer require SliceSource instances, renamed slice_obj -> slice_item, updated docs
