+++ docs/api.md
@@ -5,8 +5,8 @@ All described objects can be imported from the `zappend.api` module.
 ## Function `zappend()`
 
 ::: zappend.api.zappend
-options:
-show_root_heading: true
+    options:
+        show_root_heading: true
 
 ## Class `SliceSource`
 
@@ -27,21 +27,21 @@ show_root_heading: true
 ## Types
 
 ::: zappend.api.SliceObj
-options:
-show_root_heading: true
+    options:
+        show_root_heading: true
 
 ::: zappend.api.SliceCallable
-options:
-show_root_heading: true
+    options:
+        show_root_heading: true
 
 ::: zappend.api.ConfigItem
-options:
-show_root_heading: true
+    options:
+        show_root_heading: true
 
 ::: zappend.api.ConfigList
-options:
-show_root_heading: true
+    options:
+        show_root_heading: true
 
 ::: zappend.api.ConfigLike
-options:
-show_root_heading: true
+    options:
+        show_root_heading: true
