+++ zappend/slice/cm.py
@@ -3,10 +3,13 @@
 # https://opensource.org/licenses/MIT.
 
 import contextlib
+from typing import Any
 
 import xarray as xr
 
-from .abc import SliceSource
+from ..context import Context
+from .source import SliceSource
+from .source import to_slice_source
 
 
 class SliceSourceContextManager(contextlib.AbstractContextManager):
@@ -29,3 +32,49 @@ class SliceSourceContextManager(contextlib.AbstractContextManager):
     def __exit__(self, *exception_args):
         self._slice_source.dispose()
         self._slice_source = None
+
+
+def open_slice_dataset(
+    ctx: Context,
+    slice_item: Any,
+    slice_index: int = 0,
+) -> SliceSourceContextManager:
+    """Open the slice source for given slice object `slice_item`.
+
+    The intended and only use of the returned slice source is as context
+    manager. When used as context manager the slice source yields a slice
+    dataset.
+
+    If `slice_source` is specified in the configuration, it defines either a
+    class derived from `zappend.slice.SliceSource` or a function that returns
+    instances of `zappend.slice.SliceSource`. The retrieved slice source will
+    be returned by this function. The class or function will receive positional
+    and keyword arguments derived from `slice_item` as follows:
+
+    * `tuple`: a pair of the form `(args, kwargs)`, where `args` is a list
+      or tuple of positional arguments and `kwargs` is a dictionary of keyword
+      arguments;
+    * `list`: positional arguments only;
+    * `dict`: keyword arguments only;
+    * Any other type is interpreted as single positional argument.
+
+    If `slice_source` is not specified in the configuration, the slice object
+    `slice_item` may have one of the following types:
+
+    * `str`: A local file path or URI pointing to a dataset file such as a
+      Zarr or NetCDF. If it is a URI, the `ctx.slice_storage_options` apply.
+    * `xarray.Dataset`: An in-memory xarray dataset instance.
+    * `zappend.api.FileObj`: A file object instance pointing to a dataset
+      file in a predefined filesystem.
+    * `zappend.api.SliceSource`: A slice source. Returned as-is.
+
+    Args:
+        ctx: The processing context
+        slice_item: A slice item
+        slice_index: Optional slice index, used for dataset identification
+
+    Returns:
+        A new slice source instance
+    """
+    slice_source = to_slice_source(ctx, slice_item, slice_index)
+    return SliceSourceContextManager(slice_source)
