+++ zappend/slice/sources/__init__.py
@@ -1,0 +1,7 @@
+# Copyright © 2024 Norman Fomferra
+# Permissions are hereby granted under the terms of the MIT License:
+# https://opensource.org/licenses/MIT.
+
+from .memory import MemorySliceSource
+from .persistent import PersistentSliceSource
+from .temporary import TemporarySliceSource
